fix(init): move init into their respective modules
BREAKING CHANGE: when `require` ing `hoodie-server` that object exposed an init
method before, that could be passed a custom `env_config`. This method is gone.
You can only do `require('hoodie-server')(options, cb)` anymore, because hook
and plugin initialization now happens on the env_config object that can't be
passed from the outside.
