+++ lib/config.js
@@ -80,6 +80,10 @@ module.exports = function (options) {
     configStore.set('dbAuth', config.db.auth)
   }
 
+  if (!config.db.auth) {
+    throw new Error('Authentication details missing from database URL')
+  }
+
   if (!config.db.password) {
     config.db.password = config.db.auth.split(':')[1]
   }
